[Storage][Blob][DataLake]Add MatchConditions to Storage Packages
MatchConditions is in core package now, this commit is to move
the class to storage packages and modify that class.
