+++ sdk/storage/azure-storage-blob/azure/storage/blob/_deserialize.py
@@ -10,8 +10,9 @@ from typing import (  # pylint: disable=unused-import
     TYPE_CHECKING
 )
 
+from ._generated.models import Metrics, CorsRule, RetentionPolicy, StaticWebsite
 from ._shared.response_handlers import deserialize_metadata
-from ._models import BlobProperties, ContainerProperties
+from ._models import BlobProperties, ContainerProperties, BlobAnalyticsLogging
 
 if TYPE_CHECKING:
     from azure.storage.blob._generated.models import PageList
@@ -55,3 +56,28 @@ def get_page_ranges_result(ranges):
     if ranges.clear_range:
         clear_range = [{'start': b.start, 'end': b.end} for b in ranges.clear_range]
     return page_range, clear_range  # type: ignore
+
+
+def service_stats_deserialize(generated):
+    """Deserialize a ServiceStats objects into a dict.
+    """
+    return {
+        'geo_replication': {
+            'status': generated.geo_replication.status,
+            'last_sync_time': generated.geo_replication.last_sync_time,
+        }
+    }
+
+
+def service_properties_deserialize(generated):
+    """Deserialize a ServiceProperties objects into a dict.
+    """
+    return {
+        'analytics_logging': BlobAnalyticsLogging._from_generated(generated.logging),  # pylint: disable=protected-access
+        'hour_metrics': Metrics._from_generated(generated.hour_metrics),  # pylint: disable=protected-access
+        'minute_metrics': Metrics._from_generated(generated.minute_metrics),  # pylint: disable=protected-access
+        'cors': [CorsRule._from_generated(cors) for cors in generated.cors],  # pylint: disable=protected-access
+        'target_version': generated.default_service_version,  # pylint: disable=protected-access
+        'delete_retention_policy': RetentionPolicy._from_generated(generated.delete_retention_policy),  # pylint: disable=protected-access
+        'static_website': StaticWebsite._from_generated(generated.static_website),  # pylint: disable=protected-access
+    }
