+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/_blob_service_client_async.py
@@ -28,9 +28,8 @@ from ._container_client_async import ContainerClient
 from ._blob_client_async import BlobClient
 from .._models import (
     ContainerProperties,
-    service_stats_deserialize,
-    service_properties_deserialize,
 )
+from azure.storage.blob._deserialize import service_stats_deserialize, service_properties_deserialize
 from ._models import ContainerPropertiesPaged
 
 if TYPE_CHECKING:
@@ -442,7 +441,7 @@ class BlobServiceClient(AsyncStorageAccountHostsMixin, BlobServiceClientBase):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword ~azure.core.MatchConditions match_condition:
+        :keyword ~azure.storage.blob.MatchConditions match_condition:
             The match condition to use upon the etag.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
