+++ sdk/storage/azure-storage-blob/tests/test_append_blob_async.py
@@ -16,7 +16,7 @@ from os import (
 )
 import unittest
 
-from azure.core import MatchConditions
+from azure.storage.blob import MatchConditions
 from azure.core.exceptions import HttpResponseError, ResourceNotFoundError, ResourceModifiedError
 from azure.core.pipeline.transport import AioHttpTransport
 from multidict import CIMultiDict, CIMultiDictProxy
