+++ sdk/storage/azure-storage-blob/tests/test_page_blob.py
@@ -11,7 +11,7 @@ import pytest
 import os
 import unittest
 from datetime import datetime, timedelta
-from azure.core import MatchConditions
+from azure.storage.blob import MatchConditions
 from azure.core.exceptions import HttpResponseError, ResourceExistsError, ResourceModifiedError
 
 from azure.storage.blob import (
