+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/__init__.py
@@ -55,6 +55,7 @@ __all__ = [
     'PathPropertiesPaged',
     'LeaseProperties',
     'ContentSettings',
+    'MatchConditions'
     'AccountSasPermissions',
     'FileSystemSasPermissions',
     'DirectorySasPermissions',
