+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/_data_lake_file_client.py
@@ -215,7 +215,7 @@ class DataLakeFileClient(PathClient):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword :class:`MatchConditions` match_condition:
+        :keyword :class:~azure.storage.filedatalake.MatchConditions` match_condition:
             The match condition to use upon the etag.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
@@ -411,7 +411,7 @@ class DataLakeFileClient(PathClient):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword :class:`MatchConditions` match_condition:
+        :keyword :class:~azure.storage.filedatalake.MatchConditions` match_condition:
             The match condition to use upon the etag.
         :keyword int max_concurrency:
             The number of parallel connections with which to download.
