+++ sdk/storage/azure-storage-file-datalake/azure/storage/filedatalake/_file_system_client.py
@@ -174,7 +174,7 @@ class FileSystemClient(StorageAccountHostsMixin):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword :class:`MatchConditions` match_condition:
+        :keyword :class:~azure.storage.filedatalake.MatchConditions` match_condition:
             The match condition to use upon the etag.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
@@ -254,7 +254,7 @@ class FileSystemClient(StorageAccountHostsMixin):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword :class:`MatchConditions` match_condition:
+        :keyword :class:~azure.storage.filedatalake.MatchConditions` match_condition:
             The match condition to use upon the etag.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
@@ -328,7 +328,7 @@ class FileSystemClient(StorageAccountHostsMixin):
         :keyword str etag:
             An ETag value, or the wildcard character (*). Used to check if the resource has changed,
             and act according to the condition specified by the `match_condition` parameter.
-        :keyword :class:`MatchConditions` match_condition:
+        :keyword :class:~azure.storage.filedatalake.MatchConditions` match_condition:
             The match condition to use upon the etag.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
